Refactor. Extract earthquakes module
